requirements.txt
numpy~=1.22.1
pytesseract~=0.3.8
opencv-python~=4.5.5.62
Pillow~=9.0.0
pynput~=1.7.6